move status decorator to helper functions, moved the status tab components to analysis_builder dir, moved 404 component to helper components
